For Riff Fork development: set ver 1.0.0-dev.0; rename package to riff-mm-webapp
Setting the version to 1.0.0-dev.0 indicates the start of development on
the Riff MM version 1.0.0

Subsequent development versions will bump the prerelease number until it is
released (when it is merged into the `riff-master` branch, and the prerelease info
is removed).
At that time the develop branch version should be bumped to either 1.1.0-dev.0 or
2.0.0-dev.0 (I prefer the 1st as the 2.0.0 designation can be decided at release
time because going from 1.1.0-dev.N to 2.0.0 is easy, but going from 2.0.0-dev.N
to 1.1.0 is problematic).