replace most uses of underscore package in actions/views/dashboard.js (style errors 3114 -> 3071)
The underscore package is pretty large, and most of the uses are now available in the
ecmascript language (map, filter, sort...) so it would be nice to replace all uses
of the package so we can remove the dependency.

I also rewrote some of the algorithms slightly while I was at it, all results are
still supposed to be the same, if they're not I've got a bug.

Well except for the comparison of startDate which is a Date object, and I'm sure it
wasn't supposed to compare object but the date it represents (so I changed it to compare
startTime).