fix some prop types
I added prop types for MeetingViz, but it was my best guess from examining the code and
I got some things wrong.

I changed size.height and with to not be required and size to be not exact:

The prop `size.height` is marked as required in `MeetingViz`, but its value is `null`.
    in MeetingViz (created by Connect(MeetingViz))

Failed prop type: Invalid prop `size` key `position` supplied to `MeetingViz`.
Bad object: {
  "width": 0,
  "height": null,
  "position": null
}

Similarly:

Invalid prop `influenceData` of type `array` supplied to `MeetingViz`, expected `object`.
    in MeetingViz (created by Connect(MeetingViz))

Failed prop type: Invalid prop `influenceData` of type `object` supplied to `MeetingViz`, expected `array`.
    in MeetingViz (created by Connect(MeetingViz))
    in Connect(MeetingViz) (created by SizeMeRenderer(Connect(MeetingViz)))
    in SizeMeReferenceWrapper (created by SizeMeRenderer(Connect(MeetingViz)))
    in SizeMeRenderer(Connect(MeetingViz)) (created by SizeMe(Connect(MeetingViz)))
    in SizeMe(Connect(MeetingViz)) (created by DashboardView)
    in div (created by DashboardView)
    in div (created by DashboardView)
    in DashboardView (created by PureLifecycle(DashboardView))

This error has been happening for a while (so it was not caused by recent changes)
and is still not fixed:
TypeError: Cannot read property '_id' of undefined
    at Object.maybeLoadNextMeeting (index.js?fefa:97)
    at MeetingViz.loadThisAndMaybeMore (MeetingViz.?6296:142)