added setLogLevel method to logger; exported config object w/ pseudo config values for dataserver and signalmaster
Those config values now have defaults suitable for the expected docker service stack w/
the reverse proxy forwarding the websocket connections for the data server and signalmaster.

I left the process.env code there for the time being although that would hardcode the values
at build time. The right thing to do is to get the values from the mm-server config.

I moved getting the pseudo config values for the signalmaster url and path into the `utils/riff/index.js`
so that it is all together.