Announce when members of the chat join or leave the room
Added a function addA11yBrowserAlert(string) in utils which temporarily injects a div into the dom, which means the screen reader will read its contents(the passed string). This function is used to alert users using screen readers when other users join or leave the room.

Will need to test this with other people, since your window has to be active(focused) for the screenreader to read out these messages (I believe)