Override some MM eslint styles for Riff source files (in components/dashboard)
Where the entire tree consists of Riff source files we can add a `.eslint.json` file with
our override styles.
When the Riff sources are mixed with MM sources we have to override the styles by using
an eslint comment at the top of the file.

In general we don't want to override too many styles, but there are some things we really
prefer to do differently.