Remove gitlab CI/CD deploy jobs and supporting script and elasticbeanstalk (EBS) config files
The deploy jobs were always a little flaky and they haven't worked at all for several months.

Nobody has found a need for them, and they spin up reasonably large AWS instances which have
to be terminated manually, and cost a fair amount.

This commit removing the continuous deployment is available and hopefully can be located if
the removed files need to be examined for setting up this functionality (working) again.