fix issue w/ npm install NOT completely installing ALL packages! (in particular mattermost-redux)
The node docker image user is root
npm acts differently when run as root

The problem was that mattermost-redux was not getting properly installed.
This is because it installs from a git url not a npm package repository.
Installing from a git url, mattermost-redux depends on npm noticing that
it has a `prepare` script in its package.json, an installing its dependency
and then running that script.

However npm when run as root has unsafe-perm set to false and won't run scripts
automatically. It seems it also doesn't install dependencies in this situation.

For some reason only setting unsafe-perm to true was not sufficient. When I compared
the other config differences between the root user and a non-root user, I saw
that the root user's shell was `bash` while the non-root user's shell was `/bin/bash`
Setting the shell config to `/bin/bash` seemed to do the trick.