improve the Makefile check-style-ci target to error only if a threshold is exceeded
This allows us to fail pipelines if the style errors increase, as such allow_failure
was removed from the check-style job.

It does mean that to stay on top of fixes to existing style errors the `.gitlab-ci.yml`
file will have to be constantly edited as we work our way down to 0 style errors.

Also changes the check-style log format to html