gitlab ci can't run jobs on github pull requests, so change to run on fix, feature and meta branches
See this gitlab issue: https://gitlab.com/gitlab-org/gitlab-ee/issues/5667

If it ever gets resolved, maybe we can make this work better.
It also has some discussion about using webhooks which we might be able to use
to create the desired functionality of only running the CI on PRs and tags.

Also streamlined the gitlab-ci.yml more using a template and anchors/aliases