Implement 1st pass at a course connections graph and some cleanup of the dashboard layout
The course connections graph utilizes data from a new UserAnalytics API call to the mattermost
server. It is mostly intended to show data for courses where the users sign-in via an LTI link
from an LMS, and where there are 2 private channels the users are assigned to. These channels
are referred to as "plg" and "capstone" in the Nexted course that is the 1st user of this
feature.

- Added SectionHeader component for rendering section titles and descriptions in the dashboard
- Made sure all labels and tooltip meet accessibility requirements

Notes:
- Still needs data for a11y chart table

This commit is the squashed result after all code review changes of
Brec's original PR (some by Brec, some by Mike). The overall design of the
chart is all Brec (figuring out how to use amcharts). Mike did some refactoring
during code review to allow redraw and to streamline the code.