Show posts by the current user on the course connections chart and additional tooltips
Posts by the current user in a given context are now counted as interactions in that
context.

Added tooltips for contexts to display interaction counts
Added a tooltip for the 'You' node to display total interaction counts

The tooltips for a context node and a user node are now very similar,
so the similarities have been extracted into a separate static method.

This commit is a combination of original work by Brec w/ some code review
refactoring by Mike Lippert