add history package, fix usage of history package
The history package was being used but wasn't actually listed as a dependency.
It was working because it is a dependency of react-router and react-router-dom.
In addition it was writing an error that the current method being used to
access createBrowserHistory was deprecated.

```
Warning: Please use `require("history").createBrowserHistory` instead of `require("history/createBrowserHistory")`. Support for the latter will be removed in the next major release.
```