remove bogus ref string attributes
Chat no longer loads. We are getting this message in the console (well it leads to this message):
```
Element ref was specified as a string (local) but no owner was set. This could happen for one of the following reasons: 1. You may be adding a ref to a function component 2. You may be adding a ref to a component that was not created inside a component's render method 3. You have multiple copies of React loaded See https://fb.me/react-refs-must-have-owner for more information
```

I'm pretty sure this was caused by the ref in styled.js